% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_winP.R
\name{calculate_winP}
\alias{calculate_winP}
\title{Calculate Win Probability and Other Metrics
R package written by Pavel S Roshanov June 29 2024
Based on idea and SAS code in Zou G, Zou L, Qiu S Fang. Parametric and nonparametric methods for confidence intervals and sample size planning for win probability in parallel-group randomized trials with Likert item and Likert scale data. Pharmaceutical Statistics. 2023 May 1;22(3):418–39.}

\usage{
calculate_winP(data, group_var, post_var, pre_var = NULL)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{group_var}{A string specifying the name of the grouping variable.}

\item{post_var}{A string specifying the name of the post-baseline variable.}

\item{pre_var}{An optional string specifying the name of the pre-baseline variable (not required).}
}
\value{
A data frame containing the calculated metrics.
}
\description{
This function calculates the Win Probability (WinP) and its confidence interval
along with other related metrics.
}
